infra/docker/templates/observability/example-app/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-exporter-jaeger>=1.21.0
